Move 'return' tests from 'define' to own package.
