+++ src/test/java/tests/statements/define/DefineTest.java
@@ -106,76 +106,6 @@ public class DefineTest {
         }
 
         @Test
-        void itCanReturnWithNumber() {
-            String sourceFile = "statements/define/return-with-number.frendli";
-            String actual = run(sourceFile);
-            String expected = "123";
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnWithText() {
-            String sourceFile = "statements/define/return-with-text.frendli";
-            String actual = run(sourceFile);
-            String expected = "Some text";
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnWithBoolean() {
-            String sourceFile = "statements/define/return-with-boolean.frendli";
-            String actual = run(sourceFile);
-            String expected = "true";
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnWithExplicitEmpty() {
-            String sourceFile = "statements/define/return-with-explicit-empty.frendli";
-            String actual = run(sourceFile);
-            String expected = "empty";
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnWithImplicitEmpty() {
-            String sourceFile = "statements/define/return-with-implicit-empty.frendli";
-            String actual = run(sourceFile);
-            String expected = """
-                    empty
-                    empty
-                    """.trim();
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnWithNonLiteralExpression() {
-            String sourceFile = "statements/define/return-with-non-literal-expression.frendli";
-            String actual = run(sourceFile);
-            String expected = """
-                    102
-                    123456
-                    """.trim();
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnFromNestedBlock() {
-            String sourceFile = "statements/define/return-from-nested-block.frendli";
-            String actual = run(sourceFile);
-            String expected = "123";
-            assertEquals(expected, actual);
-        }
-
-        @Test
-        void itCanReturnFromNestedLocalFunction() {
-            String sourceFile = "statements/define/return-from-nested-local-function.frendli";
-            String actual = run(sourceFile);
-            String expected = "123";
-            assertEquals(expected, actual);
-        }
-
-        @Test
         void itCanUseBuiltInDefinitionName() {
             String sourceFile = "statements/define/use-built-in-definition-name.frendli";
             String actual = run(sourceFile);
