+++ src/test/java/tests/statements/return_/return-from-nested-local-function.frendli
